mybaynexx-main/requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
exceptiongroup==1.1.2
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==23.1
pycparser==2.21
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
wsproto==1.2.0
psycopg2==2.9.9  # Para S.O Windows.
#psycopg2-binary==2.9.9  # Para S.O Linux.
mysql-connector-python==9.0.0
oracledb==2.3.0